fix issue where picker could be invisible when first shown
There was a case where, if the picker was made first responder BEFORE
the window was even shown, it would go invisible and you'd never see the
expanded panel.

Looks like you can't setup child windows when a parent window has not
yet been displayed.

To workaround, don't try to show the expanded picker if the parent
window is not visible.