in becomeFirstResponder, only show the panel if it's not already shown
Fixes issue where, on window restoration, when AppKit goes to restore
the first responder state to whichever view had it, becomeFirstResponder
would get called twice.  So, we have to guard against that and only show
the panel if it's not already app.

In my testing, successive calls to `-[NSWindow makeFirstResponder:]` did
not result in multiple calls into becomeFirstResponder.  Maybe window
restoration is special in that way.